demo/agentcore/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arise-devops-agent"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "strands-agents >= 1.13.0",
    "strands-agents-tools >= 0.2.0",
    "arise-ai >= 0.1.0",
    "boto3 >= 1.38.0",
    "a2a-sdk[http] >= 0.3.0",
    "fastapi >= 0.115.0",
    "uvicorn >= 0.34.0",
]
